fix index.tsの中身を整理
