+++ src/client/convex/etc.ts
@@ -1,10 +1,14 @@
 import * as Discord from 'discord.js'
 import Option from 'type-of-option'
 import Settings from 'const-settings'
+import * as dateTable from '../../io/dateTable'
 import * as status from '../../io/status'
 import * as json from '../../io/json'
 import {User, Json} from '../../io/type'
 import * as util from '../../util'
+import * as over from '../../client/convex/over'
+import * as situation from '../../client/convex/situation'
+import * as plan from '../../client/plan/delete'
 
 /**
  * 同時凸の持越計算を行う
@@ -38,7 +42,7 @@ const overCalc = (HP: number, a: number, b: number): number => {
  * メッセージ送信者にタスキルロールを付与する
  * @param msg DiscordからのMessage
  */
-export const AddTaskKillRoll = async (msg: Discord.Message) => {
+export const AddTaskKillRole = async (msg: Discord.Message) => {
   // 既にタスキルしてるか確認する
   const isRole = util.IsRole(msg.member, Settings.ROLE_ID.TASK_KILL)
 
@@ -50,6 +54,47 @@ export const AddTaskKillRoll = async (msg: Discord.Message) => {
 
     msg.reply('タスキルロールを付けたわよ！')
   }
+}
+
+/**
+ * 全員のタスキルロールを外す
+ */
+export const RemoveTaskillRole = () => {
+  // 全員のメンバー一覧を取得
+  const guildMembers = util.GetGuild()?.members.cache.map(m => m)
+
+  // メンバー全員のタスキルロールを外す
+  guildMembers?.forEach(m => m?.roles.remove(Settings.ROLE_ID.TASK_KILL))
+
+  // bot-notifyに通知をする
+  const channel = util.GetTextChannel(Settings.CHANNEL_ID.BOT_NOTIFY)
+  channel.send('全員のタスキルロールを外したわ')
+
+  console.log('remove task kill role')
+}
+
+/**
+ * クラバトがある日に、クランメンバー全員に凸残ロールを付与する
+ */
+export const SetRemainConvex = async () => {
+  // クラバトの日じゃない場合は終了
+  const date = await dateTable.TakeDate()
+  if (date.num === '練習日') return
+
+  // べろばあのクランメンバー一覧を取得
+  const clanMembers = util
+    .GetGuild()
+    ?.roles.cache.get(Settings.ROLE_ID.CLAN_MEMBERS)
+    ?.members.map(m => m)
+
+  // クランメンバーに凸残ロールを付与する
+  clanMembers?.forEach(m => m?.roles.add(Settings.ROLE_ID.REMAIN_CONVEX))
+
+  // bot-notifyに通知をする
+  const channel = util.GetTextChannel(Settings.CHANNEL_ID.BOT_NOTIFY)
+  channel.send('クランメンバーに凸残ロールを付与したわ')
+
+  console.log('Add convex role')
 }
 
 /**
@@ -80,7 +125,45 @@ export const RemoveBossRole = async (member: Option<Discord.GuildMember>) => {
 }
 
 /**
- * スプレッドシートのメンバー一覧を更新する
+ * 凸予定、持越、凸残ロールを全て削除する
+ */
+export const ResetAllConvex = () => {
+  // 凸予定を全て削除
+  plan.DeleteAll()
+
+  // 持越を全て削除
+  over.AllDeleteMsg()
+
+  // 凸残ロールを全て削除
+  // べろばあのクランメンバー一覧を取得
+  const clanMembers = util
+    .GetGuild()
+    ?.roles.cache.get(Settings.ROLE_ID.CLAN_MEMBERS)
+    ?.members.map(m => m)
+
+  clanMembers?.forEach(m => m?.roles.remove(Settings.ROLE_ID.REMAIN_CONVEX))
+}
+
+/**
+ * メンバー全員の凸状況をリセットする
+ */
+export const ResetConvex = async () => {
+  // 全員の凸状況をリセット
+  await status.ResetConvex()
+
+  // メンバー全員の状態を取得
+  const members = await status.Fetch()
+  situation.Report(members)
+
+  // bot-notifyに通知をする
+  const channel = util.GetTextChannel(Settings.CHANNEL_ID.BOT_NOTIFY)
+  channel.send('全員の凸状況をリセットしたわ')
+
+  console.log('Reset convex')
+}
+
+/**
+ * botのクランメンバー一覧を更新する
  * @param msg DiscordからのMessage
  */
 export const UpdateMembers = async (msg: Discord.Message) => {
@@ -100,7 +183,18 @@ export const UpdateMembers = async (msg: Discord.Message) => {
   // ステータスを更新
   await status.UpdateUsers(users)
 
-  // jsonの値を更新
+  // jsonの値を送信
   const list: Json = users.map(u => ({[u.id]: u.name})).reduce((a, b) => ({...a, ...b}))
-  await json.Update(list)
+  await json.Send(list)
+}
+
+/**
+ * デイリーミッション消化の通知する
+ */
+export const NotifyDailyMission = () => {
+  // 雑談に通知をする
+  const channel = util.GetTextChannel(Settings.CHANNEL_ID.CHAT)
+  channel.send('もう4:30よ！デイリーミッションは消化したわよね！！してなかったらぶっ殺すわよ！！！！')
+
+  console.log('Notify daily mission digestion')
 }
