+++ tsconfig.json
@@ -64,6 +64,6 @@
       "src/util/cron.ts",
       "src/util/index.ts",
       "src/util/prototype.ts",
-      "src/util/setting.ts",
+      "src/init/setting.ts",
   ]
 }
